requirements.txt
slack_bolt>=1.18.0
slack_sdk>=3.27.0
python-dotenv>=1.0.1
pydantic>=2.7.1
cohere>=5.5.8
numpy>=1.26.4
requests>=2.32.3
fastapi>=0.111.0
uvicorn>=0.30.0
